+++ src/components/common/offCanvas/OffCanvasMenu.style.js
@@ -1,0 +1,74 @@
+import Styled, {keyframes} from 'styled-components';
+import colours from '../../../style/colours';
+import typography from '../../../style/typography';
+
+
+const slideIn = keyframes`
+    from {
+        transform: translateX(-100vw);
+    }
+    to {
+    
+    }
+`;
+
+
+const CanvasBg = Styled.div`
+    background: rgba(0,0,0,.5);
+    top: 0;
+    width: calc(100vw + 150px);          
+    height: 100vh;
+    position: absolute;
+    z-index: 100;
+    animation: ${slideIn} alternate 1s ease-in-out;
+`;
+
+
+const MobileMenuWrapper = Styled.div`
+    top: 0;
+    height: 100vh;
+    width: 150px;
+    background: ${colours.secondary};    
+    padding-top: 50px;
+    z-index: 1000;
+    position: fixed;
+    animation: ${slideIn} alternate 1s ease-in-out;
+    box-shadow: 1px 1px 9px 5px rgba(0,0,0,0.32);
+    
+    nav {
+        margin-top: 70px;
+        z-index: 1100;
+    }
+    
+    ul {
+        top: 25px;
+        max-width: 150px;
+        margin: 0;
+        padding: 0;
+        z-index: 10;
+        > li {
+            list-style-type: none;
+            font-family: ${typography.body};
+            font-size: .8em;
+            color: #fff;
+            margin: 5px 0;
+            padding: 10px 0;
+            text-align: center;
+            text-transform: uppercase;     
+            
+           > a {
+                color: #fff;
+                padding-bottom: 5px;
+                
+                &:hover {
+                    border-bottom: 2px solid ${colours.accent};
+                    color: #fff;
+            }
+                
+           }            
+        }
+    }
+    
+`;
+
+export { CanvasBg, MobileMenuWrapper };
